requirements.txt
torch==2.0.1
transformers==4.33.0
sentence-transformers==2.2.2
huggingface_hub==0.16.4
faiss-cpu==1.7.4
pandas==2.0.3
numpy==1.24.3
tqdm==4.66.1
click==8.1.7